requirements.txt
fastapi
python-whois
requests
beautifulsoup4
pyOpenSSL
uvicorn
streamlit
pandas
plotly
reportlab
schedule
dnspython